+++ docs/project/changelog.rst
@@ -117,6 +117,9 @@ Bug fixes
 
 * Restored compatibility of the ``websockets`` CLI with Windows.
 
+* Fixed a bug that could delay or block the client in the :mod:`threading`
+  implementation on macOS when the opening handshake fails.
+
 .. _16.1.1:
 
 16.1.1